==> R/02_calc_indices.Rmd <==
---
title: "Calculate indices"
output: html_notebook
---


```{r}
library(tidyverse)
```



```{r}
data <- read_csv("../data/data_analysis.csv")
plabryatt <- read_csv("../data/PLABRYATT_v4.csv")
spp_lookup <- read_csv("../data/spp_lookup.csv")
```

quick check if plabryatt species list works
```{r}
in_plabryatt <- spp_lookup$species_nbn %in% plabryatt$`PLABRYATT_VTA NAME`
table(spp_lookup$species_nbn[which(spp_lookup$taxon_grp == "vascular plant")] %in% plabryatt$`PLABRYATT_VTA NAME`)
table(spp_lookup$taxon_grp[!in_plabryatt])
```

pain in the xxx as usual....


what to do...

ok, so I matched PLABRYATT with gibif using taxize and same for species_comparison


```{r}
plabryatt <- read_csv("../data/PLABRYATT_v5.csv")
```
<<<<<<< HEAD

```{r}
in_plabryatt <- spp_lookup$gbif.taxon_id %in% plabryatt$gbif.taxon_id
table(spp_lookup$gbif.taxon_id[which(spp_lookup$taxon_grp == "vascular plant")] %in% plabryatt$gbif.taxon_id)
table(spp_lookup$taxon_grp[!in_plabryatt])
```

...seems to have got worse.  OK, need to lookup accepted names from GBIF taxon_id
=======

```{r}
in_plabryatt <- spp_lookup$gbif.taxon_id %in% plabryatt$gbif.taxon_id
table(spp_lookup$gbif.taxon_id[which(spp_lookup$taxon_grp == "vascular plant")] %in% plabryatt$gbif.taxon_id)
table(spp_lookup$taxon_grp[!in_plabryatt])
```

...seems to have got worse.  OK, need to lookup accepted names from GBIF taxon_id



... still struggling to automate this.  Instead did manual matching using MAVIS helper doc Preparing_data_for_MAVIS.xls

```{r}
spp_lookup_MAVIS <- readxl::read_xls("../data/Preparing_data_for_MAVIS_stodmarsh.xls", sheet = 1)
spp_lookup_MAVIS

```

```{r}
spp_lookup <- spp_lookup %>% left_join(select(spp_lookup_MAVIS, species_verbatim, species_MAVIS))
#write_csv(spp_lookup, "../data/spp_lookup.csv")
```

```{r}
spp_data <- read_csv("../data/data_analysis.csv")
```

```{r}
spp_lookup <- read_csv("../data/spp_lookup.csv")
```

```{r}
spp_data <- spp_data %>%   left_join(select(spp_lookup, species_verbatim, species_MAVIS), by = c("spp" = "species_verbatim"))

```

check we've got all species in the lookup
```{r}
spp_data %>% select(spp, species_MAVIS) %>% 
  filter(is.na(species_MAVIS))
  # filter(species_MAVIS == "NA")
  #  filter(!species_MAVIS %in% spp_lookup$species_MAVIS)
```

... still struggling to automate this.  Instead did manual matching using MAVIS helper doc Preparing_data_for_MAVIS.xls

```{r}
spp_lookup_MAVIS <- readxl::read_xls("../data/Preparing_data_for_MAVIS_stodmarsh.xls", sheet = 1)
spp_lookup_MAVIS
```

```{r}
spp_lookup <- spp_lookup %>% left_join(select(spp_lookup_MAVIS, species_verbatim, species_MAVIS))
#write_csv(spp_lookup, "../data/spp_lookup.csv")
```

```{r}
spp_data <- read_csv("../data/data_analysis.csv")
```

```{r}
spp_lookup <- read_csv("../data/spp_lookup.csv")
```

```{r}
spp_data <- spp_data %>%   left_join(select(spp_lookup, species_verbatim, species_MAVIS), by = c("spp" = "species_verbatim"))

```

#### format for MAVIS without abundance
```{r}
data_for_MAVIS_bsln <- spp_data %>% 
  filter(survey == "bsln") %>% 
  select(ditch_id, species_MAVIS)

data_for_MAVIS_plt <- spp_data %>% 
  filter(survey == "plt") %>% 
  select(ditch_id, species_MAVIS)

#write_csv(data_for_MAVIS_bsln, "../data/data_for_MAVIS_stodm_bsln.csv")
#write_csv(data_for_MAVIS_plt, "../data/data_for_MAVIS_stodm_plt.csv")
```


#### format for MAVIS with abundance
```{r}
data_for_MAVIS_abnd_bsln <- spp_data %>% 
  filter(survey == "bsln") %>% 
  select(ditch_id, species_MAVIS, abund)

data_for_MAVIS_abnd_plt <- spp_data %>% 
  filter(survey == "plt") %>% 
  select(ditch_id, species_MAVIS, abund)

#write_csv(data_for_MAVIS_bsln, "../data/data_for_MAVIS_stodm_bsln.csv")
#write_csv(data_for_MAVIS_plt, "../data/data_for_MAVIS_stodm_plt.csv")
```

**Baseline and pilot used differetn measures of abundance.** 

Baseline = *Dominance recorded as  1 = 1-9%  2 = 10-59%  3 = 60-100%*

Pilot = *DAFOR, or 1 = present, unless scrub or Phramites australis 1 =1-9%, 2= 10-59%, 3=60-100%*


```{r}
hist(data_for_MAVIS_abnd_bsln$abund)
hist(data_for_MAVIS_abnd_plt$abund)
```


```{r}
count(data_for_MAVIS_abnd_plt, abund)
unique(data_for_MAVIS_abnd_plt$abund)
```
